fix: move to nominatim.ts the logic to extract a territory, so it is now reusable for AddressAutocomplete and TerritoryFilter components
